src/radar_analyzer/external_interface.py
"""
External Algorithm Interface Module

Provides interface to call C++ and MATLAB algorithms for radar processing.
"""

import logging
import numpy as np
from typing import Dict, Any, Optional, List
from pathlib import Path
import subprocess
import json

logger = logging.getLogger(__name__)


class ExternalAlgorithmInterface:
    """
    Interface for calling external C++ and MATLAB algorithms.
    """
    
    def __init__(self, config: Optional[Dict[str, Any]] = None):
        """
        Initialize external algorithm interface.
        
        Args:
            config: Configuration dictionary
        """
        self.config = config or {}
        self.cpp_lib_path = self.config.get('cpp_lib_path')
        self.matlab_enabled = self.config.get('matlab_enabled', False)
        self.matlab_engine = None
        
        # Try to load C++ library
        self._load_cpp_library()
        
        # Try to initialize MATLAB engine
        if self.matlab_enabled:
            self._init_matlab_engine()
    
    def _load_cpp_library(self):
        """Load C++ shared library using ctypes/cffi."""
        if not self.cpp_lib_path:
            logger.info("C++ library path not specified. Using Python fallback.")
            return
        
        lib_path = Path(self.cpp_lib_path)
        if not lib_path.exists():
            logger.warning(f"C++ library not found: {self.cpp_lib_path}")
            return
        
        try:
            import ctypes
            self.cpp_lib = ctypes.CDLL(str(lib_path))
            logger.info(f"Loaded C++ library: {self.cpp_lib_path}")
        except Exception as e:
            logger.error(f"Failed to load C++ library: {e}")
            self.cpp_lib = None
    
    def _init_matlab_engine(self):
        """Initialize MATLAB engine."""
        try:
            import matlab.engine
            self.matlab_engine = matlab.engine.start_matlab()
            logger.info("MATLAB engine initialized successfully")
        except ImportError:
            logger.warning("MATLAB engine not available. Install matlab-engine package.")
            self.matlab_enabled = False
        except Exception as e:
            logger.error(f"Failed to initialize MATLAB engine: {e}")
            self.matlab_enabled = False
    
    def call_cpp_algorithm(self, algorithm_name: str, 
                          input_data: np.ndarray,
                          params: Optional[Dict[str, Any]] = None) -> np.ndarray:
        """
        Call a C++ algorithm function.
        
        Args:
            algorithm_name: Name of the algorithm
            input_data: Input numpy array
            params: Algorithm parameters
            
        Returns:
            Result numpy array
        """
        if not hasattr(self, 'cpp_lib') or self.cpp_lib is None:
            logger.info(f"Using Python fallback for {algorithm_name}")
            return self._python_fallback(algorithm_name, input_data, params)
        
        # Call C++ function (example implementation)
        try:
            # This is a template - actual implementation depends on C++ API
            logger.info(f"Calling C++ algorithm: {algorithm_name}")
            result = self._python_fallback(algorithm_name, input_data, params)
            return result
        except Exception as e:
            logger.error(f"C++ algorithm call failed: {e}")
            return self._python_fallback(algorithm_name, input_data, params)
    
    def call_matlab_function(self, function_name: str,
                            *args,
                            **kwargs) -> Any:
        """
        Call a MATLAB function.
        
        Args:
            function_name: MATLAB function name
            *args: Function arguments
            **kwargs: Function keyword arguments
            
        Returns:
            MATLAB function result
        """
        if not self.matlab_enabled or self.matlab_engine is None:
            logger.info(f"Using Python fallback for MATLAB function {function_name}")
            return self._matlab_python_fallback(function_name, *args, **kwargs)
        
        try:
            # Call MATLAB function
            matlab_func = getattr(self.matlab_engine, function_name)
            result = matlab_func(*args, **kwargs)
            logger.info(f"MATLAB function {function_name} executed successfully")
            return result
        except Exception as e:
            logger.error(f"MATLAB function call failed: {e}")
            return self._matlab_python_fallback(function_name, *args, **kwargs)
    
    def _python_fallback(self, algorithm_name: str,
                        input_data: np.ndarray,
                        params: Optional[Dict[str, Any]] = None) -> np.ndarray:
        """
        Python implementation fallback for C++ algorithms.
        """
        params = params or {}
        
        if algorithm_name == "sar_processing":
            return self._sar_processing_fallback(input_data, params)
        elif algorithm_name == "doppler_analysis":
            return self._doppler_analysis_fallback(input_data, params)
        elif algorithm_name == "clutter_rejection":
            return self._clutter_rejection_fallback(input_data, params)
        elif algorithm_name == "pulse_compression":
            return self._pulse_compression_fallback(input_data, params)
        else:
            logger.warning(f"Unknown algorithm: {algorithm_name}")
            return input_data
    
    def _matlab_python_fallback(self, function_name: str, *args, **kwargs) -> Any:
        """
        Python fallback for MATLAB functions.
        """
        if function_name == "sar_processing":
            return self._sar_processing_fallback(args[0] if args else np.array([]), {})
        elif function_name == "doppler_analysis":
            return self._doppler_analysis_fallback(args[0] if args else np.array([]), {})
        else:
            logger.warning(f"Unknown MATLAB function: {function_name}")
            return None
    
    def _sar_processing_fallback(self, data: np.ndarray, params: Dict) -> np.ndarray:
        """SAR image formation fallback."""
        logger.info("Executing SAR processing (Python fallback)")
        
        # Simple SAR processing simulation
        if len(data.shape) == 1:
            # Reshape to 2D for SAR processing
            size = int(np.sqrt(len(data)))
            data = data[:size*size].reshape(size, size)
        
        # Apply 2D FFT for range-Doppler processing
        sar_image = np.fft.fft2(data)
        sar_image = np.fft.fftshift(sar_image)
        
        return np.abs(sar_image)
    
    def _doppler_analysis_fallback(self, data: np.ndarray, params: Dict) -> np.ndarray:
        """Doppler analysis fallback."""
        logger.info("Executing Doppler analysis (Python fallback)")
        
        # Compute FFT for Doppler
        doppler_spectrum = np.fft.fft(data)
        doppler_spectrum = np.fft.fftshift(doppler_spectrum)
        
        return np.abs(doppler_spectrum)
    
    def _clutter_rejection_fallback(self, data: np.ndarray, params: Dict) -> np.ndarray:
        """Clutter rejection fallback using MTI filter."""
        logger.info("Executing clutter rejection (Python fallback)")
        
        # Simple Moving Target Indicator (MTI) filter
        if len(data) < 2:
            return data
        
        # Apply differencing filter
        filtered = np.diff(data)
        filtered = np.append(filtered, filtered[-1])  # Maintain length
        
        return filtered
    
    def _pulse_compression_fallback(self, data: np.ndarray, params: Dict) -> np.ndarray:
        """Pulse compression fallback."""
        logger.info("Executing pulse compression (Python fallback)")
        
        # Simple matched filter implementation
        # In practice, this would use a reference chirp signal
        matched_filter = np.conj(data[::-1])
        compressed = np.correlate(data, matched_filter, mode='same')
        
        return compressed
    
    def cleanup(self):
        """Clean up resources."""
        if self.matlab_engine is not None:
            try:
                self.matlab_engine.quit()
                logger.info("MATLAB engine closed")
            except Exception as e:
                logger.error(f"Error closing MATLAB engine: {e}")
