requirements.txt
# Radar Target Behavior Analysis System - Dependencies

# Core data processing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# GUI Framework
PyQt5>=5.15.9
matplotlib>=3.7.0

# Data I/O and format handling
h5py>=3.8.0
pillow>=9.5.0

# C++/MATLAB Integration
cffi>=1.15.1
matlab-engine>=9.13.0  # Optional, install separately if MATLAB is available

# Machine Learning (for feature extraction and classification)
scikit-learn>=1.2.0

# Configuration and logging
pyyaml>=6.0
colorlog>=6.7.0

# Testing
pytest>=7.3.0
pytest-cov>=4.0.0

# Synthetic data generation
noise>=1.2.2  # Perlin noise for realistic trajectories
